requirements.txt
flask>=3.0.0
flask-restx>=1.3.0
google-genai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
Werkzeug>=3.0.0
